+++ starsky/starsky/clientapp/src/containers/media-content.tsx
@@ -4,7 +4,7 @@ import Notification, {
 } from "../components/atoms/notification/notification";
 import Preloader from "../components/atoms/preloader/preloader";
 import HealthStatusError from "../components/molecules/health-status-error/health-status-error";
-import MenuDefault from "../components/organisms/menu-default/menu-default";
+import ApplicationException from "../components/organisms/application-exception/application-exception";
 import ArchiveContextWrapper from "../contexts-wrappers/archive-wrapper";
 import DetailViewContextWrapper from "../contexts-wrappers/detailview-wrapper";
 import useSockets from "../hooks/realtime/use-sockets";
@@ -32,14 +32,7 @@ const MediaContent: React.FC = () => {
 
   const settings = useGlobalSettings();
   const language = new Language(settings.language);
-  const MessageApplicationException = language.text(
-    "We hebben een op dit moment een verstoring op de applicatie",
-    "We have a disruption on the application right now"
-  );
-  const MessageRefreshPageTryAgain = language.text(
-    "Herlaad de pagina om het opnieuw te proberen",
-    "Please reload the page to try again"
-  );
+
   const MessageConnectionRealtimeError = language.text(
     "De verbinding is niet helemaal oké. We proberen het te herstellen",
     "The connection is not quite right. We are trying to fix it"
@@ -74,11 +67,7 @@ const MediaContent: React.FC = () => {
       ) : null}
       {pageType === PageType.Unauthorized ? <Login /> : null}
       {pageType === PageType.ApplicationException ? (
-        <>
-          <MenuDefault isEnabled={false} />
-          <div className="content--header">{MessageApplicationException}</div>
-          <div className="content--subheader">{MessageRefreshPageTryAgain}</div>
-        </>
+        <ApplicationException></ApplicationException>
       ) : null}
       {pageType === PageType.Archive &&
       archive &&
